requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
matplotlib==3.3.0
numpy==1.18.5
pandas==1.0.5
Pillow==7.1.2
scikit-learn==0.24.1
torch==1.5.1
torchvision==0.6.1
tqdm==4.46.1
xlrd==1.2.0